+++ CHANGELOG.md
@@ -1,1 +1,3 @@
-- Dailies checked -alydev
+- Valentiones 2026 instance support, huge thanks to Jaksuhn and the rest of the Punish team -alydev
+  - ok it works now -alydev
+- Changed "don't skip cutscenes" checkbox to also turn off text skip -alydev